Struct boa_datagen::PluralRangesFallbackHack

pub(crate) struct PluralRangesFallbackHack(pub(crate) DatagenProvider);

Hack that associates the und locale with an empty plural ranges data. This enables the default behaviour for all locales without data.

Tuple Fields
0: DatagenProvider